% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-package.R
\docType{package}
\name{agecrypt-package}
\alias{agecrypt}
\alias{agecrypt-package}
\title{agecrypt: File Encryption with the age Format}
\description{
A dependency-free implementation of the
\href{https://age-encryption.org/v1}{age} v1 file encryption format,
backed by a vendored copy of the C implementation \code{agec}. Encrypts to
one or more X25519 recipients or with a passphrase, with optional ASCII
armor. SSH-key recipients and plugins are not supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedrobtz/agecrypt}
  \item Report bugs at \url{https://github.com/pedrobtz/agecrypt/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item agec authors (Authors of the vendored 'agec' C implementation, https://git.sr.ht/~min/agec) [contributor, copyright holder]
}

}
\keyword{internal}
